board 수정(2)
