+++ src/components/Board/BoardItem/BoardItem.jsx
@@ -4,11 +4,12 @@ import {GoComment , GoEye} from 'react-icons/go';
 import { Link } from 'react-router-dom';
 
 export function BoardItem ({item,color,value}){
+    const shortItemTItle = item.title.slice(0,14);
     return (
         <Link to={`./${value}/${item.id}`}>
             <RightDivItem >
                 <Animal size="100px" color={color} circle dance/>
-                <RightItemTitle>{item.title}</RightItemTitle>
+                <RightItemTitle>{item.title.length <= 14 ? item.title : `${shortItemTItle} ...`}</RightItemTitle>
                 <RightItemSeparate>
                     <RightItemComments><GoComment style={{marginRight:"3px"}}/>{item.comments}</RightItemComments>
                     <RightItemViews><GoEye style={{marginRight:"3px"}}/>{item.views}</RightItemViews>
@@ -19,11 +20,12 @@ export function BoardItem ({item,color,value}){
 }
 
 export function SoloBoardItem ({item,color}){
+    const shortItemTItle = item.title.slice(0,14);
     return (
         <Link to={`./${item.id}`}>
             <RightDivItem >
                 <Animal size="100px" color={color} circle dance/>
-                <RightItemTitle>{item.title}</RightItemTitle>
+                <RightItemTitle>{item.title.length <= 14 ? item.title : `${shortItemTItle} ...`}</RightItemTitle>
                 <RightItemSeparate>
                     <RightItemComments><GoComment style={{marginRight:"3px"}}/>{item.comments}</RightItemComments>
                     <RightItemViews><GoEye style={{marginRight:"3px"}}/>{item.views}</RightItemViews>
